fix : interface in component moved to type file
