+++ src/types/type.ts
@@ -37,4 +37,22 @@ export interface ChatModalProp {
   title: string | undefined;
 }
 
+export interface ListData {
+  id: number;
+  listName: string;
+  value: string;
+}
+
+export interface PagenatedData {
+  account_type: string;
+  email: string;
+  id: number;
+  mbti: string;
+  name: string;
+  nickname: string;
+  status: string;
+  thumbnail_url: string;
+  user_type: string;
+}
+
 export type idProp = number | undefined;
